+++ engine/baml-schema-wasm/Cargo.toml
@@ -11,7 +11,10 @@ license-file.workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [lints.rust]
+dead_code = "deny"
 elided_named_lifetimes = "deny"
+unused_imports = "deny"
+unused_variables = "deny"
 
 [dependencies]
 reqwest.workspace = true
